+++ src/exports/extensions/attachment.ts
@@ -407,7 +407,7 @@ export const Attachment = Node.create<AttachmentOptions>({
       },
       // When it's .attachment, its coming from <action-text-attachment><figure></figure></action-text-attachment> its the raw HTML.
       {
-        tag: "action-text-attachment > figure.attachment",
+        tag: "action-text-attachment:not([content]) > figure.attachment",
         contentElement: "figcaption",
         getAttrs: (node) => {
           const isValid = canParseAttachment(node, this.options.previewable);
@@ -420,7 +420,7 @@ export const Attachment = Node.create<AttachmentOptions>({
         },
       },
       {
-        tag: "action-text-attachment",
+        tag: "action-text-attachment[content]",
         getAttrs: (node) => {
           const isValid = canParseAttachment(node, this.options.previewable);
 
@@ -528,7 +528,7 @@ export const Attachment = Node.create<AttachmentOptions>({
       attachmentId: { default: null },
       altTextDialogOpen: { default: false },
       caption: {
-        default: "",
+        default: null,
         parseHTML: (element) => {
           return (
             element.querySelector("figcaption")?.innerHTML ||
@@ -564,39 +564,39 @@ export const Attachment = Node.create<AttachmentOptions>({
         },
       },
       sgid: {
-        default: "",
+        default: null,
         parseHTML: (element) => findAttribute(element, "sgid"),
       },
       src: {
-        default: "",
+        default: null,
         parseHTML: (element) => findAttribute(element, "src"),
       },
       height: {
-        default: "",
+        default: null,
         parseHTML: (element) => findAttribute(element, "height"),
       },
       width: {
-        default: "",
+        default: null,
         parseHTML: (element) => {
           return findAttribute(element, "width");
         },
       },
       contentType: {
-        default: "",
+        default: null,
         parseHTML: (element) => {
           return parseContentTypeFromElement(element);
         },
       },
       fileName: {
-        default: "",
+        default: null,
         parseHTML: (element) => findAttribute(element, "filename"),
       },
       fileSize: {
-        default: "",
+        default: null,
         parseHTML: (element) => findAttribute(element, "filesize"),
       },
       content: {
-        default: "",
+        default: null,
         parseHTML: (element) => {
           const attachment = element.closest("action-text-attachment");
 
@@ -625,7 +625,7 @@ export const Attachment = Node.create<AttachmentOptions>({
         },
       },
       url: {
-        default: "",
+        default: null,
         parseHTML: (element) => {
           return findAttribute(element, "url");
         },
@@ -714,13 +714,18 @@ export const Attachment = Node.create<AttachmentOptions>({
 
           const { tr } = view.state;
 
-          const captionNode = view.state.doc.nodeAt(getPos() + 1);
+          const pos = getPos();
+          if (pos == null) {
+            return;
+          }
+
+          const captionNode = view.state.doc.nodeAt(pos + 1);
           captionNode?.nodeSize;
 
           tr.setSelection(
             TextSelection.create(
               view.state.doc,
-              getPos() + 1 + (captionNode?.nodeSize || 0),
+              pos + 1 + (captionNode?.nodeSize || 0),
             ),
           );
 
@@ -762,10 +767,14 @@ export const Attachment = Node.create<AttachmentOptions>({
 
       const handleMouseMove = (_e: MouseEvent) => {
         if (mouseIsDown && typeof getPos === "function") {
+          const pos = getPos();
+          if (pos == null) {
+            return;
+          }
           const { view } = editor;
           view.dispatch(
             view.state.tr.setSelection(
-              NodeSelection.create(view.state.doc, getPos()),
+              NodeSelection.create(view.state.doc, pos),
             ),
           );
         }
@@ -783,6 +792,9 @@ export const Attachment = Node.create<AttachmentOptions>({
           const { tr } = view.state;
 
           const pos = getPos();
+          if (pos == null) {
+            return;
+          }
           tr.setNodeMarkup(pos, null, {
             ...node.attrs,
             ...attrs,
@@ -798,6 +810,9 @@ export const Attachment = Node.create<AttachmentOptions>({
           const { tr } = view.state;
 
           const pos = getPos();
+          if (pos == null) {
+            return;
+          }
           tr.delete(pos, pos + 1);
           view.dispatch(tr);
         }